+++ package/bin/input_module_webex_meetings_summary_report.py
@@ -1,15 +1,14 @@
 import json
 from datetime import datetime, timedelta, timezone
-from dateutil.relativedelta import *
 
 from webex_constants import (
     _MEETING_USAGE_REPORTS_ENDPOINT,
     _MEETING_ATTENDEE_REPORTS_ENDPOINT,
-    _RESPONSE_TAG_MAP,
-    _TOKEN_EXPIRES_CHECKPOINT_KEY,
+    _RESPONSE_TAG_MAP
 )
 from webex_api_client import paging_get_request_to_webex
-from oauth_helper import update_access_token
+from oauth_helper import get_valid_access_token
+from webex_utils import get_time_span
 
 '''
     IMPORTANT
@@ -35,35 +34,9 @@ def collect_events(helper, ew):
     account_name = opt_global_account.get("name")
     client_id = opt_global_account.get("client_id")
     client_secret = opt_global_account.get("client_secret")
-    access_token = opt_global_account.get("access_token")
-    refresh_token = opt_global_account.get("refresh_token")
+    stored_access_token = opt_global_account.get("access_token")
+    stored_refresh_token = opt_global_account.get("refresh_token")
     base_endpoint = opt_global_account.get("endpoint")
-
-    # check if the access token expired
-    # get the access_token_expired_time checkpoint
-    expiration_checkpoint_key = _TOKEN_EXPIRES_CHECKPOINT_KEY.format(
-        account_name=account_name
-    )
-    access_token_expired_time = helper.get_check_point(expiration_checkpoint_key)
-
-    now = datetime.now(timezone.utc)
-
-    # update the access token if it expired
-    if (
-        not access_token_expired_time
-        or datetime.strptime(access_token_expired_time, "%m/%d/%Y %H:%M:%S").replace(tzinfo=timezone.utc) < now
-    ):
-
-        helper.log_debug(
-            "[*] The access token of account {account_name} expired! Updating now!".format(
-                account_name=account_name
-            )
-        )
-
-        # override the access_token and expires_in
-        access_token, refresh_token, expires_in = update_access_token(
-            helper, account_name, client_id, client_secret, refresh_token, base_endpoint
-        )
     
     # check the checkpoint
     # get startdate from checkpoint
@@ -71,45 +44,29 @@ def collect_events(helper, ew):
         helper.get_input_stanza_names()
     )
     timestamp = helper.get_check_point(last_timestamp_checkpoint_key)
-    #timestamp = None
     helper.log_debug("[-] last time timestamp: {}".format(timestamp))
-
-    # set up start time
-    # first time start_time from UI
-    if timestamp is None:
-        start_time = opt_start_time
-        # save the UI start_time as checkpoint
-        helper.save_check_point(
-            last_timestamp_checkpoint_key, start_time
-        )
-        helper.log_debug("[-] no checkpoint timestamp exists, saving new timestamp: {}".format(start_time))
-
-    else:
-        # shift 1 second to avoid duplicate
-        start_time = (
-            datetime.strptime(timestamp, "%Y-%m-%dT%H:%M:%SZ") + timedelta(seconds=1)
-        ).strftime("%Y-%m-%dT%H:%M:%SZ")
-
-    # set up end time
-    now = datetime.now(timezone.utc)
-    # add 24hr delay
-    now_delay = now - timedelta(hours=24)
-
+    
+    # End time is has a different validation for this input
+    start_time, _ = get_time_span(opt_start_time, opt_end_time, timestamp, "%Y-%m-%dT%H:%M:%SZ")
+    
+    # 24 hours delay
+    now_delay = datetime.now(timezone.utc) - timedelta(hours=24)
+    
     if opt_end_time and datetime.strptime(opt_end_time, "%Y-%m-%dT%H:%M:%SZ").replace(tzinfo=timezone.utc) < now_delay:
         end_time = opt_end_time
     else:
         end_time = now_delay.strftime("%Y-%m-%dT%H:%M:%SZ")
-
-    # compare if start_time ?> end_time, if so, break
-    if datetime.strptime(start_time, "%Y-%m-%dT%H:%M:%SZ") > datetime.strptime(
-        end_time, "%Y-%m-%dT%H:%M:%SZ"
-    ):
+    
+    # if start time is not returned it means it has completed the ingestion
+    if not start_time:
         helper.log_info(
             "[-] Finished ingestion for time range {start_time} - {end_time}".format(
-                start_time=start_time, end_time=end_time
+                start_time=opt_start_time, end_time=opt_end_time
             )
         )
         return
+    
+    access_token, refresh_token = get_valid_access_token(helper, account_name, client_id, client_secret, stored_access_token, stored_refresh_token, base_endpoint)
 
     # construct the request params for meetings endpoint
     meeting_usage_reports_params = {"siteUrl": opt_site_url}
@@ -136,13 +93,14 @@ def collect_events(helper, ew):
     # write meetings data into Splunk
 
     try:
-        checkpoint_time = datetime.strptime(helper.get_check_point(last_timestamp_checkpoint_key), "%Y-%m-%dT%H:%M:%SZ")
+        checkpoint_value = helper.get_check_point(last_timestamp_checkpoint_key) or opt_start_time
+        checkpoint_time = datetime.strptime(checkpoint_value,"%Y-%m-%dT%H:%M:%SZ")
+        
         for meeting in meeting_usage_reports:
             # compare the meeting start time with the last checkpoint time
-            last_checkpoint_time = datetime.strptime(
-                helper.get_check_point(last_timestamp_checkpoint_key),
-                "%Y-%m-%dT%H:%M:%SZ",
-            )
+            last_checkpoint_value = helper.get_check_point(last_timestamp_checkpoint_key) or opt_start_time
+            last_checkpoint_time = datetime.strptime(last_checkpoint_value,"%Y-%m-%dT%H:%M:%SZ")
+            
             meeting_start_time = datetime.strptime(meeting["start"], "%Y-%m-%dT%H:%M:%SZ")
             # ingest the meetings that happened after the last ingestion
             if meeting_start_time > last_checkpoint_time: