+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v1.3.2] - 2026-02-24
+
+### Changed
+
+- Refactored input file logic to use helper functions for token validation and time handling.
+- Standardized all date formats across all the inputs.
+- Updated documentation to reflect the date format standardization.
+
 ## [v1.3.1] - 2025-12-27
 
 ### Fixed